Changed type of pointer, and added null and nil messages

+++ ft_printf.h
@@ -19,13 +19,13 @@
 int		ft_printf(char const *format, ...);
 int		interpret_conversion(char const *format, va_list *args);
 
-char	write_base16(unsigned long long n, char start);
+char	write_base16(unsigned long n, char start);
 int		ft_putstr_fd(char *s, int fd);
 int		ft_putchar_fd(char c, int fd);
 
-int		ft_putnbr_base16_fd(unsigned long long n, int fd, char base);
+int		ft_putnbr_base16_fd(unsigned long n, int fd, char base);
 int		ft_putnbr_unsigned_fd(unsigned int n, int fd);
 int		ft_putnbr_fd(int n, int fd);
 
-int		ft_putptr_fd(unsigned long long n, int fd);
+int		ft_putptr_fd(unsigned long n, int fd);
 #endif
